FIXING TENSORFLOW AND BUG
Bug was found inside main.py in which the directions left and right were swapped, though the algorithm before pretraining could've solved this issue. Tensorflow pretrain was also fixed to limit speed and hopefully lead to a successful end goal. Next addition is to modify the genetic.py class to set a limited range to which is was trained.